+++ pcweb/pages/security/views/__init__.py
@@ -1,0 +1,11 @@
+"""Views package initialization."""
+
+from .header import security_title
+from .grid import security_grid
+from .features_table import features_table_section
+
+__all__ = [
+    "security_title",
+    "security_grid",
+    "features_table_section"
+]
